+++ packages/client/src/common/LoadingMessage/loadingMessageStyles.ts
@@ -1,0 +1,23 @@
+/*
+ * © 2021 Thoughtworks, Inc.
+ */
+
+import { makeStyles } from '@material-ui/core/styles'
+
+const PADDING_LOADING = 2
+
+const useStyles = makeStyles((theme) => ({
+  loadingContainer: {
+    flexDirection: 'column',
+    alignItems: 'center',
+    justifyContent: 'center',
+    justifyItems: 'center',
+    minHeight: '100vh',
+  },
+  loadingMessage: {
+    padding: theme.spacing(PADDING_LOADING),
+    fontSize: '24px',
+  },
+}))
+
+export default useStyles