+++ packages/core/src/unknown/UnknownUsage.ts
@@ -1,0 +1,10 @@
+/*
+ * © 2021 Thoughtworks, Inc.
+ */
+
+import { IUsageData } from '../.'
+
+export default interface UnknownUsage extends IUsageData {
+  readonly cost: number
+  readonly usageUnit: string
+}