+++ packages/api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cloud-carbon-footprint/api
 
+## 1.1.0
+
+### Minor Changes
+
+- 52a8b3c1: Adds support for specifying recommendation target for AWS via API parameter
+
+  Please refer to this [commit](https://github.com/cloud-carbon-footprint/cloud-carbon-footprint/commit/404c77796ae838766dc8007e18daee2c7526f6ed) to update create app templates.
+
+### Patch Changes
+
+- Updated dependencies [f75cf08f]
+- Updated dependencies [52a8b3c1]
+  - @cloud-carbon-footprint/app@0.2.0
+  - @cloud-carbon-footprint/common@1.1.0
+
 ## 1.0.0
 
 ### Major Changes